Change database solution from CockroachDB to YugabyteDB (Rejected)
This commit rejects the proposal to move from CockroachDB to YugabyteDB as the chosen database solution. The decision was made by V-Sekai and fire. The rejection is based on potential new problems and challenges associated with adopting a new database solution, as well as the possible learning curve for the team in working with YugabyteDB. Additionally, it was noted that online primary key upgrades cannot be done with YugabyteDB.